// src/newOrder.dto.ts
export class NewOrderDto{
  tej: string;
  nev: string;
  szaml_Orszag: string;
  szaml_IranySzam: number;
  szaml_Varos: string;
  szaml_UtcaHsz: string;
  szall_Orszag: string;
  szall_IranySzam: number;
  szall_Varos: string;
  szall_UtcaHsz: string;
  kartyaSzam: string;
  lej_datum: Date;
  cvv: number;
  kuponkod: string;
}